Pull out ActiveLaunchProfileValueProviderBase
Pull out a new abstract type, `ActiveLaunchProfileValueProviderBase`, to serve as the base type for `IInterceptingPropertyValueProvider`s that need to read or write properties within the active launch profile. `ExecutablePathValueProvider` and `LaunchTargetValueProvider` have been updated to make use of this new type.
